Add user registry
Create the model for the user with Sequelize and sync it
with the database. Add functionality to the userController for account registration and confirmation. Add the render for signup,
login and forgot password in the userController file.
Add express-validator to validate the form through the registration
process.
Add bcrypt for hashing and salt the passwords within the model.
Add the Helpers folder and add the emails file with the module
of Nodemailer to register the email, and pass it the key value
pairs intended for the email. Add to the Helpers folder a file
called Tokens to generate a random token to send via email
for account confirmation, and delete it after the user
confirms the email request. Send the error messages and render
them in the views.
Add to the views folder the field for authorization
and adds the views for registration, login, forgot-password and
confirmation. And in another folder for Templates the view
for messages to send between the registration process
and before the confirmation.
Also add the folder and file for the routes.