Add section for forgotten password
The function resetPassword is in charge of the post request.
Add the result of the validationResult request and see if the user does
not exist in the database. Copy the error from the Register module
to the resetPassword module.
Generate a new token for the forgotten password.
Create a new email form in the helpers folder, and send the password
reset email. Then show the message of this action.
Check if the token sent to the user exists in the database and if it
is related to the specific user that requested the new password.
Rethrow the error if the user does not exist.
Render the form for the password reset field.
Add the same validation of express-validator for the email, check that
the result is not empty, identify the resetting user with the one-time
validation token and hash their new password with bcrypt,
then empty the token field and save the user.
Render a new confirmation view with the message
of the new password saved successfully.