Add user authentication
Add the function to authenticate the user in the userController.
Add the POST method and action to the form in the login view.
Add below the form element in the login view the hidden type input, name and token value we generated earlier.
Generate the token for each render response.
Import the same verification fields for email and password from the registration form, and change the messages to the same required fields.
First check if the user's email already exists in the database. Otherwise, display the validation error messages to the user. Second, check if the user's email is confirmed in the database, if not, render the same validation errors above, but with the message that the user hasn't confirmed his account yet.
Use bcrypt to compare passwords with the compareSync function. To do this, add a function to compare passwords in the prototype of the user model, which returns the result of the operation. In the controller, check if the entered password corresponds to the password in the database. If it's false, it returns the validation errors as before, with the corresponding error message.
Add jsonwebtoken and generate a new token that expires after one day for each time the user authenticates, store the token in a cookie with cookie-parser, and then redirect the user to the new view.